refactor: 세션보드 엔티티를 common 모듈로 이동
단일 DB 환경에서 엔티티를 팀 모듈별로 분리할 때 발생하는 구조적 모순을 해소하기 위함.

- 팀 간 공유가 필요한 엔티티(User, Generation 등)는 결국 어딘가에 한 번 정의되어야 함
- 단일 DataSource 위에서 동작하므로 영속성 컨텍스트는 어차피 하나
- FK 타입 계약(UUID vs Integer 등)이 팀 사이를 떠돌면 PR마다 동기화 비용 발생
- common에 모으면 동일 DB 스키마에 대한 단일 진실 공급원(SSOT) 확보

엔티티 15개의 패키지를 com.study.sessionboard.entity → com.study.common.entity 로 이동.
참조하는 코드 없음(import 0건), session-board는 이미 common을 의존하므로 빌드에 영향 없음.
