+++ common/src/main/java/com/study/common/entity/Comment.java
@@ -1,4 +1,4 @@
-package com.study.sessionboard.entity;
+package com.study.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -17,23 +17,28 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
-@Table(name = "session_note")
+@Table(name = "comment")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SessionNote {
+public class Comment {
 
   @Id
   @GeneratedValue(strategy = GenerationType.UUID)
   private UUID id;
 
   @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "session_id", nullable = false)
-  private Session session;
+  @JoinColumn(name = "post_id", nullable = false)
+  private EventPost post;
 
   @Column(name = "author_id", nullable = false)
   private UUID authorId;
 
-  private String body;
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "parent_id")
+  private Comment parent;
+
+  @Column(nullable = false)
+  private String content;
 
   @Column(name = "created_at", nullable = false, updatable = false)
   private OffsetDateTime createdAt;
@@ -43,10 +48,17 @@ public class SessionNote {
     createdAt = OffsetDateTime.now();
   }
 
-  public static SessionNote of(Session session, UUID authorId) {
-    SessionNote note = new SessionNote();
-    note.session = session;
-    note.authorId = authorId;
-    return note;
+  public static Comment of(EventPost post, UUID authorId, String content) {
+    Comment comment = new Comment();
+    comment.post = post;
+    comment.authorId = authorId;
+    comment.content = content;
+    return comment;
+  }
+
+  public static Comment ofReply(EventPost post, UUID authorId, Comment parent, String content) {
+    Comment comment = of(post, authorId, content);
+    comment.parent = parent;
+    return comment;
   }
 }
