+++ common/src/main/java/com/study/common/entity/ResourceVisibility.java
@@ -1,4 +1,4 @@
-package com.study.sessionboard.entity;
+package com.study.common.entity;
 
 public enum ResourceVisibility {
   PUBLIC,
