+++ common/src/main/java/com/study/common/entity/SessionStatus.java
@@ -1,4 +1,4 @@
-package com.study.sessionboard.entity;
+package com.study.common.entity;
 
 public enum SessionStatus {
   scheduled,
